create test package
