+++ test_autolens/point/analysis/test_all_to_all.py
@@ -1,31 +1,6 @@
 import pytest
 
 from autolens.point.analysis import AnalysisAllToAllPointSource
-import autolens as al
-import autofit as af
-
-
-@pytest.fixture
-def grid():
-    return al.Grid2D.uniform(
-        shape_native=(100, 100),
-        pixel_scales=0.05,
-    )
-
-
-@pytest.fixture
-def model():
-    lens = af.Model(
-        al.Galaxy,
-        redshift=0.5,
-        mass=al.mp.Isothermal,
-    )
-    source = af.Model(al.ps.PointSourceChi)
-
-    return af.Collection(
-        lens=lens,
-        source=source,
-    )
 
 
 def test_likelihood(grid, model):
